refactor: [#31] sidedish 컴포넌트에서 card, tabMenu 분리
