--- frontend/sidedish/src/SmallSidedish.css
@@ -1,118 +1,0 @@
-small-sidedish {
-    display: flex;
-    flex-direction: column;
-    border-bottom: 1px solid #ebebeb;
-    padding: 56px 0px;
-}
-.small-sidedish__header {
-    border-bottom: 1px solid #ebebeb;
-}
-.small-sidedish__title {
-    font-size: 2rem;
-    font-weight: 700;
-    line-height: 58px;
-    letter-spacing: -0.004em;
-    text-align: left;
-    margin: 0 auto;
-    width: 1280px;
-}
-.small-sidedish__tab-menu {
-    display: flex;
-    gap: 32px;
-    margin: 24px auto 0;
-    width: 1280px;
-}
-.small-sidedish__menu-item {
-    font-size: 2rem;
-    font-weight: 500;
-    line-height: 30px;
-    letter-spacing: -0.008em;
-    text-align: center;
-    padding-bottom: 15px;
-    cursor: pointer;
-}
-.small-sidedish__cards {
-    display: flex;
-    margin-top: 34px;
-    margin: 0 auto;
-    width: 1280px;
-}
-.current-tab {
-    border-bottom: solid 2px black;
-}
-.small-sidedish__cards {
-    display: grid;
-    grid-template-columns: repeat(3, 1fr);
-    margin-top: 34px;
-}
-.small-sidedish__card-img-container {
-    width: 300px;
-    height: 300px;
-}
-.small-sidedish__card-img {
-    width: 100%;
-    height: 100%;
-    object-fit: conver;
-}
-.small-sidedish__card-item--text {
-    display: flex;
-    flex-direction: column;
-    gap: 8px;
-    margin-top: 16px;
-    margin-bottom: 22px;
-}
-.small-sidedish__card-name {
-    font-weight: 500;
-    font-size: 1.6rem;
-    line-height: 26px;
-    letter-spacing: -0.008em;
-    color: #3f3f3f;
-}
-.small-sidedish__card-description {
-    font-weight: 400;
-    font-size: 1.4rem;
-    line-height: 24px;
-    letter-spacing: -0.008em;
-    color: #777;
-}
-.small-sidedish__card-prices {
-    display: flex;
-    gap: 8px;
-}
-.big-sidedish__card-price--default {
-    font-weight: 500;
-    font-size: 1.6rem;
-    line-height: 26px;
-    letter-spacing: -0.008em;
-    color: #1b1b1b;
-}
-.small-sidedish__card-price--option {
-    font-weight: 400;
-    font-size: 1.4rem;
-    line-height: 24px;
-    letter-spacing: -0.008em;
-    text-decoration-line: line-through;
-    color: #bcbcbc;
-}
-.small-sidedish__card-tag--launch {
-    font-weight: 500;
-    font-size: 1.2rem;
-    line-height: 18px;
-    letter-spacing: -0.008em;
-    color: #fff;
-    background-color: #ff8e14;
-    padding: 6px 8px;
-    border-radius: 25px;
-    margin-top: 22px;
-}
-.small-sidedish__card-tag--event {
-    font-weight: 500;
-    font-size: 1.2rem;
-    line-height: 18px;
-    letter-spacing: -0.008em;
-    color: #fff;
-    background-color: #6dd028;
-    padding: 6px 8px;
-    border-radius: 25px;
-    margin-top: 22px;
-}
