+++ frontend/sidedish/src/config.js
@@ -2,4 +2,5 @@ const BASE_URL = process.env.REACT_APP_BASE_URL;
 
 export const API = {
     FESTIVAL: `${BASE_URL}/api/categories?festival=`,
+    DISH_SECTION: `${BASE_URL}/api/dish?section=`,
 };
