+++ frontend/sidedish/src/view/GlobalStyle.jsx
