+++ frontend/sidedish/src/view/sidedish/card/Card.style.js
@@ -1,63 +1,17 @@
 import styled from "styled-components";
 
-const Container = styled.div`
-    display: flex;
-    flex-direction: column;
-    border-bottom: 1px solid #ebebeb;
-    padding: 56px 0px;
+const ImgWrapper = styled.div`
+    width: 100%
+    height: 100%;
 `;
 
-const Header = styled.div`
-    border-bottom: 1px solid #ebebeb;
-`;
-
-const Title = styled.h2`
-    font-size: 4rem;
-    font-weight: 700;
-    line-height: 58px;
-    letter-spacing: -0.004em;
-    text-align: left;
-    margin: 0 auto;
-    width: 1280px;
-`;
-
-const TabBar = styled.ul`
-    display: flex;
-    gap: 32px;
-    margin: 24px auto 0;
-    width: 1280px;
-`;
-
-const Tab = styled.li`
-    font-size: 2rem;
-    font-weight: 500;
-    line-height: 30px;
-    letter-spacing: -0.008em;
-    text-align: center;
-    padding-bottom: 15px;
-    cursor: pointer;
-    border-bottom: ${(props) => (props.isCurrTab ? "solid 2px black" : "")};
-`;
-
-const CardsContainer = styled.ul`
-    display: grid;
-    grid-template-columns: repeat(3, 1fr);
-    margin: 34px auto;
-    width: 1280px;
-`;
-
-const CardImgWrapper = styled.div`
-    width: 300px;
-    height: 300px;
-`;
-
-const CardImg = styled.img`
+const Img = styled.img`
     width: 100%;
     height: 100%;
     object-fit: cover;
 `;
 
-const CardTextContainer = styled.div`
+const TextContainer = styled.div`
     display: flex;
     flex-direction: column;
     gap: 8px;
@@ -65,7 +19,7 @@ const CardTextContainer = styled.div`
     margin-bottom: 22px;
 `;
 
-const CardTitle = styled.h3`
+const Title = styled.h3`
     font-weight: 500;
     font-size: 1.6rem;
     line-height: 26px;
@@ -73,7 +27,7 @@ const CardTitle = styled.h3`
     color: #3f3f3f;
 `;
 
-const CardDescription = styled.p`
+const Description = styled.p`
     font-weight: 400;
     font-size: 1.4rem;
     line-height: 24px;
@@ -115,17 +69,11 @@ const EventBadge = styled.span`
 `;
 
 export {
-    Container,
-    Header,
+    ImgWrapper,
+    Img,
+    TextContainer,
     Title,
-    TabBar,
-    Tab,
-    CardsContainer,
-    CardImgWrapper,
-    CardImg,
-    CardTextContainer,
-    CardTitle,
-    CardDescription,
+    Description,
     PriceContainer,
     Price,
     EventBadge,
